feat(stats): show correct seller stats
